attendance-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
pydantic-settings==2.3.3
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography==42.0.7
pgvector==0.2.4
numpy==1.26.4
torch==2.2.2
torchvision==0.17.2
facenet-pytorch==2.6.0
opencv-python-headless==4.9.0.80
Pillow==10.2.0
python-multipart==0.0.9
structlog==24.2.0
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
faker==25.8.0

